polymarket-bot/requirements.txt
# Polymarket Copy Trading Bot - Dependencies

# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP & WebSocket
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Task Queue
celery[redis]==5.3.4
redis==5.0.1

# API Framework (for dashboard)
fastapi==0.105.0
uvicorn[standard]==0.25.0

# Crypto & Web3
web3==6.13.0
eth-account>=0.13.0
py-clob-client==0.22.0

# Data & Analysis
pandas==2.1.4
numpy==1.26.2

# Utilities
python-telegram-bot==20.7
structlog==23.2.0
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code Quality
ruff==0.1.8
black==23.12.1
mypy==1.7.1
